tests/wpt/web-platform-tests/soft-navigation-heuristics/dropped-entries.tentative.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<meta name="timeout" content="long">
<title>Detect simple soft navigation.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/soft-navigation-helper.js"></script>
</head>
<body>
  <main id=main>
  <a id=link>Click me!</a>
  </main>
  <script>
    const link = document.getElementById("link");

    testSoftNavigation({
      addContent: async () => {
        await addImageToMain();
      },
      link: link,
      clicks: 52,
      extraValidations: (entries, options)=>{
        assert_equals(options['droppedEntriesCount'], 2);
      },
      test: "Test that a soft navigation is detected when the click is done "
        + "on a custom element."});
  </script>
</body>
</html>
